fix bound variable
