+++ .pre-commit-config.yaml
@@ -107,6 +107,12 @@ repos:
             pytest,
             cryptography,
             bs4,
+            aiomcache,
+            hypothesis,
+            tortoise-orm,
+            mako,
+            httpx,
+            types-beautifulsoup4,
           ]
   - repo: https://github.com/ariebovenberg/slotscheck
     rev: v0.14.1
@@ -137,7 +143,7 @@ repos:
             aiomcache,
           ]
   - repo: https://github.com/pycqa/pylint
-    rev: "v2.15.4"
+    rev: "v2.15.5"
     hooks:
       - id: pylint
         exclude: "test_*"
@@ -161,7 +167,7 @@ repos:
     rev: "v0.982"
     hooks:
       - id: mypy
-        exclude: "starlette"
+        args: [--show-traceback, --pdb]
         additional_dependencies:
           [
             orjson,
@@ -181,6 +187,13 @@ repos:
             types-redis,
             redis,
             pytest,
+            aiomcache,
+            hypothesis,
+            tortoise-orm,
+            mako,
+            httpx,
+            types-beautifulsoup4,
+            cryptography,
           ]
   - repo: https://github.com/RobertCraigie/pyright-python
     rev: v1.1.276
