+++ mypy.ini
@@ -30,6 +30,8 @@ warn_untyped_fields = True
 [mypy-picologging.*]
 ignore_missing_imports = True
 
-
 [mypy-brotli.*]
+ignore_missing_imports = True
+
+[mypy-mako.*]
 ignore_missing_imports = True
