+++ tests/handlers/websocket/test_kwarg_handling.py
@@ -28,7 +28,7 @@ def test_handle_websocket_params_parsing() -> None:
     client = create_test_client(route_handlers=websocket_handler)
 
     # Set cookies on the client to avoid warnings about per-request cookies.
-    client.cookies = {"cookie": "yum"}
+    client.cookies = {"cookie": "yum"}  # type: ignore
 
     with client.websocket_connect("/1?qp=1", headers={"some-header": "abc"}) as ws:
         ws.send_json({"data": "123"})
