+++ tests/kwargs/test_layered_params.py
@@ -50,7 +50,7 @@ def test_layered_parameters_injected_correctly() -> None:
         },
     ) as client:
         # Set cookies on the client to avoid warnings about per-request cookies.
-        client.cookies = {"app4": "jeronimo"}
+        client.cookies = {"app4": "jeronimo"}  # type: ignore
 
         query = {"controller1": "99", "controller3": "tuna", "router1": "albatross", "app2": ["x", "y"]}
         headers = {"router3": "10"}
@@ -99,7 +99,7 @@ def test_layered_parameters_validation(parameter: str) -> None:
             query.pop(parameter)
 
         # Set cookies on the client to avoid warnings about per-request cookies.
-        client.cookies = cookies
+        client.cookies = cookies  # type: ignore
 
         response = client.get("/router/controller/1", params=query, headers=headers)
