+++ tests/response/test_streaming_response.py
@@ -1,4 +1,4 @@
-"""The tests in this file were adapted from:
+"""A large part of the tests in this file were adapted from:
 
 https://github.com/encode/starlette/blob/master/tests/test_responses.py And are
 meant to ensure our compatibility with their API.
